pyproject.toml
[project]
name = "jarvis-server"
version = "0.1.0"
description = "Jarvis Server — REST API, dreaming engine, git ops"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.34",
    "pydantic-settings>=2.7",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "arq>=0.26",
    "redis>=5.0",
    "httpx>=0.28",
    "structlog>=24.4",
    "openai>=1.60",
    "pyyaml>=6.0",
    "croniter>=1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx",
    "ruff>=0.9",
    "mypy>=1.0",
    "types-PyYAML>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]
